refactor: 버튼 공통 컴포넌트 사이즈 수정 및 폴더 이동
