+++ src/app/(workspace)/[teamId]/(edit)/layout.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import Header from "@/layouts/Header/Header";
+import SimpleLayout from "@/layouts/SimpleLayout";
+
+export default function TeamEditLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      <Header />
+      <SimpleLayout>{children}</SimpleLayout>
+    </>
+  );
+}
